vision/pyproject.toml
[project]
name = "vision"
version = "0.2.0"
description = "Air instrument system — hand tracking + phone fretboard + MIDI audio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe",
    "opencv-python",
    "numpy",
    "websockets",
    "pyfluidsynth",
]
